Reproduce similar problem as meeting
A code which calls an initiated logger will run.
A code which calls a non initiated logger will cause an Exception.
Unlike before when it was OK. The operator is breaking the code.
This belongs to a part of code we cannot modify ourselves in our scenario.

Now is the time to find a solution to make an initiated logger be able to receive logs from a part of the application that cannot initiate the logger.